Make __url event test more reliable
When can-route is added to the main CanJS test suite, the `__url event` regularly fails. After running this test locally, I noticed that it either completely fails if it’s run on its own, or sometimes fails because it depends on an assertion running on a timeout after a bunch of other timeouts, e.g. the holy grail of test flakiness.

Justin refactored this test for 4.0, so I’m stealing his changes with some minor modifications.